refactor(addon): create addon registry
closes #91
